// client/src/clients/deal-system-client-definitions.ts
export interface IUpdatableListItemDto {
    deleted: boolean;
    updated: boolean;
    id?: number | undefined;
}
// ...
export class UpdatableOfString implements IUpdatableOfString {
    updated!: boolean;
    value!: string;

    constructor(data?: IUpdatableOfString) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        makeObservable(this, {
            updated: observable,
            value: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.updated = _data["updated"];
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): UpdatableOfString {
        data = typeof data === 'object' ? data : {};
        let result = new UpdatableOfString();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["updated"] = this.updated;
        data["value"] = this.value;
        return data; 
    }

    clone(): UpdatableOfString {
        const json = this.toJSON();
        let result = new UpdatableOfString();
        result.init(json);
        return result;
    }
}

export interface IUpdatableOfString {
    updated: boolean;
    value: string;
}
// ...
export interface IUpdatableOfBoolean {
    updated: boolean;
    value: boolean;
}

export class UpdatableOfNullableInteger implements IUpdatableOfNullableInteger {
    updated!: boolean;
    value?: number | undefined;

    constructor(data?: IUpdatableOfNullableInteger) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        makeObservable(this, {
            updated: observable,
            value: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.updated = _data["updated"];
            this.value = _data["value"];
        }
    }

    static fromJS(data: any): UpdatableOfNullableInteger {
        data = typeof data === 'object' ? data : {};
        let result = new UpdatableOfNullableInteger();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["updated"] = this.updated;
        data["value"] = this.value;
        return data; 
    }

    clone(): UpdatableOfNullableInteger {
        const json = this.toJSON();
        let result = new UpdatableOfNullableInteger();
        result.init(json);
        return result;
    }
}

export interface IUpdatableOfNullableInteger {
    updated: boolean;
    value?: number | undefined;
}
// ...
export class LookupRequest implements ILookupRequest {
    id!: number;
    name!: string;
    description!: string;
    active?: boolean | undefined;

    constructor(data?: ILookupRequest) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
        makeObservable(this, {
            id: observable,
            name: observable,
            description: observable,
            active: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.active = _data["active"];
        }
    }

    static fromJS(data: any): LookupRequest {
        data = typeof data === 'object' ? data : {};
        let result = new LookupRequest();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        data["active"] = this.active;
        return data; 
    }

    clone(): LookupRequest {
        const json = this.toJSON();
        let result = new LookupRequest();
        result.init(json);
        return result;
    }
}

export interface ILookupRequest {
    id: number;
    name: string;
    description: string;
    active?: boolean | undefined;
}
// ...
export interface IUpdatableOfInteger {
    updated: boolean;
    value: number;
}
// ...
export class UserDto implements IUserDto {
    id?: number | undefined;
    username!: UpdatableOfString;
    name!: UpdatableOfString;
    userRoleId!: UpdatableOfNullableInteger;
    active!: UpdatableOfBoolean;
    workflowRolesInUser!: UserInWorkflowRoleDto[];
    integrationData!: UserIntegrationDataDto[];

    constructor(data?: IUserDto) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
            this.username = data.username && !(<any>data.username).toJSON ? new UpdatableOfString(data.username) : <UpdatableOfString>this.username; 
            this.name = data.name && !(<any>data.name).toJSON ? new UpdatableOfString(data.name) : <UpdatableOfString>this.name; 
            this.userRoleId = data.userRoleId && !(<any>data.userRoleId).toJSON ? new UpdatableOfNullableInteger(data.userRoleId) : <UpdatableOfNullableInteger>this.userRoleId; 
            this.active = data.active && !(<any>data.active).toJSON ? new UpdatableOfBoolean(data.active) : <UpdatableOfBoolean>this.active; 
        }
        if (!data) {
            this.username = new UpdatableOfString();
            this.name = new UpdatableOfString();
            this.userRoleId = new UpdatableOfNullableInteger();
            this.active = new UpdatableOfBoolean();
            this.workflowRolesInUser = [];
            this.integrationData = [];
        }
        makeObservable(this, {
            id: observable,
            username: observable,
            name: observable,
            userRoleId: observable,
            active: observable,
            workflowRolesInUser: observable,
            integrationData: observable,
        })
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.username = _data["username"] ? UpdatableOfString.fromJS(_data["username"]) : new UpdatableOfString();
            this.name = _data["name"] ? UpdatableOfString.fromJS(_data["name"]) : new UpdatableOfString();
            this.userRoleId = _data["userRoleId"] ? UpdatableOfNullableInteger.fromJS(_data["userRoleId"]) : new UpdatableOfNullableInteger();
            this.active = _data["active"] ? UpdatableOfBoolean.fromJS(_data["active"]) : new UpdatableOfBoolean();
            if (Array.isArray(_data["workflowRolesInUser"])) {
                this.workflowRolesInUser = [] as any;
                for (let item of _data["workflowRolesInUser"])
                    this.workflowRolesInUser!.push(UserInWorkflowRoleDto.fromJS(item));
            }
            if (Array.isArray(_data["integrationData"])) {
                this.integrationData = [] as any;
                for (let item of _data["integrationData"])
                    this.integrationData!.push(UserIntegrationDataDto.fromJS(item));
            }
        }
    }

    static fromJS(data: any): UserDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["username"] = this.username ? this.username.toJSON() : <any>undefined;
        data["name"] = this.name ? this.name.toJSON() : <any>undefined;
        data["userRoleId"] = this.userRoleId ? this.userRoleId.toJSON() : <any>undefined;
        data["active"] = this.active ? this.active.toJSON() : <any>undefined;
        if (Array.isArray(this.workflowRolesInUser)) {
            data["workflowRolesInUser"] = [];
            for (let item of this.workflowRolesInUser)
                data["workflowRolesInUser"].push(item.toJSON());
        }
        if (Array.isArray(this.integrationData)) {
            data["integrationData"] = [];
            for (let item of this.integrationData)
                data["integrationData"].push(item.toJSON());
        }
        return data; 
    }

    clone(): UserDto {
        const json = this.toJSON();
        let result = new UserDto();
        result.init(json);
        return result;
    }
}

export interface IUserDto {
    id?: number | undefined;
    username: IUpdatableOfString;
    name: IUpdatableOfString;
    userRoleId: IUpdatableOfNullableInteger;
    active: IUpdatableOfBoolean;
    workflowRolesInUser: UserInWorkflowRoleDto[];
    integrationData: UserIntegrationDataDto[];
}

export class UserInWorkflowRoleDto extends UpdatableListItemDto implements IUserInWorkflowRoleDto {
    userId!: UpdatableOfInteger;
    userName!: string;
    workflowRoleId!: UpdatableOfInteger;
    workflowRoleName!: string;
    active!: UpdatableOfBoolean;

    constructor(data?: IUserInWorkflowRoleDto) {
        super(data);
        if (data) {
            this.userId = data.userId && !(<any>data.userId).toJSON ? new UpdatableOfInteger(data.userId) : <UpdatableOfInteger>this.userId; 
            this.workflowRoleId = data.workflowRoleId && !(<any>data.workflowRoleId).toJSON ? new UpdatableOfInteger(data.workflowRoleId) : <UpdatableOfInteger>this.workflowRoleId; 
            this.active = data.active && !(<any>data.active).toJSON ? new UpdatableOfBoolean(data.active) : <UpdatableOfBoolean>this.active; 
        }
        if (!data) {
            this.userId = new UpdatableOfInteger();
            this.workflowRoleId = new UpdatableOfInteger();
            this.active = new UpdatableOfBoolean();
        }
        makeObservable(this, {
            userId: observable,
            userName: observable,
            workflowRoleId: observable,
            workflowRoleName: observable,
            active: observable,
        })
    }

    init(_data?: any) {
        super.init(_data);
        if (_data) {
            this.userId = _data["userId"] ? UpdatableOfInteger.fromJS(_data["userId"]) : new UpdatableOfInteger();
            this.userName = _data["userName"];
            this.workflowRoleId = _data["workflowRoleId"] ? UpdatableOfInteger.fromJS(_data["workflowRoleId"]) : new UpdatableOfInteger();
            this.workflowRoleName = _data["workflowRoleName"];
            this.active = _data["active"] ? UpdatableOfBoolean.fromJS(_data["active"]) : new UpdatableOfBoolean();
        }
    }

    static fromJS(data: any): UserInWorkflowRoleDto {
        data = typeof data === 'object' ? data : {};
        let result = new UserInWorkflowRoleDto();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["userId"] = this.userId ? this.userId.toJSON() : <any>undefined;
        data["userName"] = this.userName;
        data["workflowRoleId"] = this.workflowRoleId ? this.workflowRoleId.toJSON() : <any>undefined;
        data["workflowRoleName"] = this.workflowRoleName;
        data["active"] = this.active ? this.active.toJSON() : <any>undefined;
        super.toJSON(data);
        return data; 
    }

    clone(): UserInWorkflowRoleDto {
        const json = this.toJSON();
        let result = new UserInWorkflowRoleDto();
        result.init(json);
        return result;
    }
}

export interface IUserInWorkflowRoleDto extends IUpdatableListItemDto {
    userId: IUpdatableOfInteger;
    userName: string;
    workflowRoleId: IUpdatableOfInteger;
    workflowRoleName: string;
    active: IUpdatableOfBoolean;
}